refactor: Moved findItemName and findTargetItem away from presentation
They need to be reused in the decorator strategies
